# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: C100 System API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /secretary/meetings:
    get:
      operationId: listMeetingRecords
      tags: [Secretary]
      summary: List meeting records
      responses:
        "200":
          description: Meeting records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingRecord"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createMeetingRecord
      tags: [Secretary]
      summary: Create a meeting record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRecordCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRecord"
  /secretary/meetings/{id}:
    get:
      operationId: getMeetingRecord
      tags: [Secretary]
      summary: Get a meeting record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Meeting record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRecord"
    patch:
      operationId: patchMeetingRecord
      tags: [Secretary]
      summary: Update a meeting record (not allowed when status is approved)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRecordCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRecord"
        "409":
          description: Conflict — record is approved, use /revise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /secretary/meetings/{id}/approve:
    post:
      operationId: approveMeetingRecord
      tags: [Secretary]
      summary: Approve a meeting record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRecord"
  /secretary/meetings/{id}/revise:
    post:
      operationId: reviseMeetingRecord
      tags: [Secretary]
      summary: Add a revision to an approved meeting record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRecordReviseBody"
      responses:
        "200":
          description: Revised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRecord"
  /secretary/correspondence:
    get:
      operationId: listCorrespondence
      tags: [Secretary]
      summary: List correspondence log
      responses:
        "200":
          description: Correspondence entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorrespondenceEntry"
    post:
      operationId: createCorrespondenceEntry
      tags: [Secretary]
      summary: Log a correspondence entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrespondenceEntryCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrespondenceEntry"
  /secretary/correspondence/{id}:
    put:
      operationId: updateCorrespondenceEntry
      tags: [Secretary]
      summary: Update a correspondence entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrespondenceEntryCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrespondenceEntry"

  # ── Treasurer ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
...
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MeetingRecord:
      type: object
      required: [id, meetingType, title, meetingDate, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        meetingType:
          type: string
        title:
          type: string
        meetingDate:
          type: string
        status:
          type: string
        agendaText:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MeetingRecordCreate:
      type: object
      required: [meetingType, title, meetingDate]
      properties:
        meetingType:
          type: string
        title:
          type: string
        meetingDate:
          type: string
        agendaText:
          type: [string, 'null']
        notes:
          type: [string, 'null']
    CorrespondenceEntry:
      type: object
      required: [id, direction, correspondent, subject, dateSent, createdAt, updatedAt]
      properties:
        id:
          type: integer
        direction:
          type: string
        correspondent:
          type: string
        subject:
          type: string
        dateSent:
          type: string
        description:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CorrespondenceEntryCreate:
      type: object
      required: [direction, correspondent, subject, dateSent]
      properties:
        direction:
          type: string
        correspondent:
          type: string
        subject:
          type: string
        dateSent:
          type: string
        description:
          type: [string, 'null']

    # ── Treasurer schemas ────────────────────────────────────────────────────
...
    MeetingRecordReviseBody:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        agendaText:
          type: [string, 'null']
        notes:
          type: [string, 'null']
...