# lib/api-zod/src/generated/types/correspondenceEntryCreate.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * C100 System API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CorrespondenceEntryCreate {
  direction: string;
  correspondent: string;
  subject: string;
  dateSent: string;
  /** @nullable */
  description?: string | null;
}